Title: 배열의 평균, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 231227/배열의 평균/ave-of-array.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 int main() {
     int arr[2][4];
-    int sum=0, a=0,b=0;
+    double sum=0, a=0,b=0;
     for(int i=0;i<2;i++){
         for(int j=0;j<4;j++){
             cin >> arr[i][j];
